+++ src/org/elixir_lang/Elixir.bnf
@@ -251,7 +251,7 @@ unqualifiedNoParenthesesManyArgumentsCall ::= noParenthesesManyArgumentsUnqualif
                                                 implements = [
                                                   // MUST be first so visitElement is used in ElixirVisitor
                                                   "com.intellij.psi.PsiElement"
-                                                  "org.elixir_lang.psi.Call"
+                                                  "org.elixir_lang.psi.call.Call"
                                                   "org.elixir_lang.psi.Quotable"
                                                   "org.elixir_lang.psi.QuotableArguments"
                                                 ]
