+++ src/org/elixir_lang/psi/AtUnqualifiedNoParenthesesCall.java
@@ -1,5 +1,7 @@
 package org.elixir_lang.psi;
 
+import org.elixir_lang.psi.call.Call;
+
 /**
  * atPrefixOperator IDENTIFIER CALL bracketArguments
  */
