+++ src/org/elixir_lang/psi/QualifiedNoArgumentsCall.java
@@ -1,5 +1,7 @@
 package org.elixir_lang.psi;
 
+import org.elixir_lang.psi.call.Call;
+
 /**
  * <expression> dotInfixOperator relativeIdentifier !CALL
  */
