+++ src/org/elixir_lang/psi/UnqualifiedNoParenthesesCall.java
@@ -1,5 +1,6 @@
 package org.elixir_lang.psi;
 
+import org.elixir_lang.psi.call.Call;
 import org.jetbrains.annotations.NotNull;
 
 /**
