+++ src/org/elixir_lang/psi/call/Call.java
@@ -1,6 +1,7 @@
-package org.elixir_lang.psi;
+package org.elixir_lang.psi.call;
 
 import com.intellij.lang.ASTNode;
+import org.elixir_lang.psi.ElixirDoBlock;
 import org.jetbrains.annotations.Nullable;
 
 /**
