+++ src/org/elixir_lang/psi/impl/ElixirPsiImplUtil.java
@@ -12,6 +12,7 @@ import org.apache.commons.lang.NotImplementedException;
 import org.elixir_lang.ElixirLanguage;
 import org.elixir_lang.Macro;
 import org.elixir_lang.psi.*;
+import org.elixir_lang.psi.call.Call;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
